JWT authorization integrated

+++ Backend/controller/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../Models/goalModel')
+const User = require('../Models/userModel')
 
 // @desc Get goals
 // @route  GET /api/goals
@@ -7,7 +8,7 @@ const Goal = require('../Models/goalModel')
 
 
 const getGoals = asyncHandler (  async (req , res) =>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user:req.user.id})
 
     res.status(200).json(goals)
 
@@ -24,7 +25,8 @@ const createGoals =  asyncHandler (  async (req , res) =>{
     }
     
     const goal = await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user:req.user.id
     })
 
     res.status(200).json(goal)
@@ -36,11 +38,31 @@ const createGoals =  asyncHandler (  async (req , res) =>{
 // @access Private
 
 const updateGoals = asyncHandler (  async (req , res) =>{
+    // checking whether the goal exists
     const goal = await Goal.findById(req.params.id)
+
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
     }
+
+
+    // checking whether the user exist -- > cross checkig ==> in the user database
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //checking the user is the owner of the goal
+    if(goal.user.toString() != user.id){
+        res.status(401)
+        throw new Error('unauthorized access')
+    }
+
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id ,req.body,{new:true})
     res.status(200).json(updatedGoal)
 })
@@ -52,6 +74,7 @@ const updateGoals = asyncHandler (  async (req , res) =>{
 // @access Private
 
 const deleteGoals = asyncHandler (  async (req , res) =>{
+    // checking whether the goal exists
     const goal = await Goal.findById(req.params.id)
 
     if(!goal){
@@ -59,7 +82,22 @@ const deleteGoals = asyncHandler (  async (req , res) =>{
         throw new Error('Goal not found')
     }
 
-    await goal.deleteOne(); // Use deleteOne instead of remove
+    // checking whether the user exist -- > cross checkig ==> in the user database
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //checking the user is the owner of the goal
+    if(goal.user.toString() != user.id){
+        res.status(401)
+        throw new Error('unauthorized access')
+    }
+
+    await goal.deleteOne(); // Use deleteOne instead of remove() --> new version update
 
     res.status(200).json({
         id :req.params.id
